refactor: split CLI into modules
