+++ CHANGELOG.md
@@ -48,6 +48,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Store commands reorganized under `branch` and `blob` subcommands.
 - Simplified file ingestion using `anybytes::Bytes::map_file` and removed
   the `memmap2` dependency.
+- Split CLI command groups into modules under `src/cli`.
 ### Removed
 - Completed work entries have been trimmed from `INVENTORY.md` now that they are
   tracked here.
